Dwave_Demo/requirements.txt
dwave-ocean-sdk>=4.2.0
cython==0.29.22
jinja2>=3.0.0
osmnx>=1.1.2
folium==0.12.1
seaborn==0.11.1
scipy>=1.6.0
streamlit==1.9.2
streamlit-folium==0.3.0